Put seed script in package.json and moved all seeding function to sep file
