+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "react-dev": "webpack -d --watch",
-    "server-dev": "nodemon server/index.js"
+    "server-dev": "nodemon server/index.js",
+    "seed": "node db/seed.js"
   },
   "author": "Will Hudson",
   "license": "ISC",
